backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
python-socketio==5.15.0
requests==2.32.5
python-dotenv==1.0.1
beautifulsoup4==4.12.3
polyline==2.0.3
geopy==2.4.1

